+++ CHANGELOG.md
@@ -24,6 +24,7 @@
   ### Fixed
   - Fix menu accordion ovelapping [#369](https://github.com/cartoway/planner-web/pull/369)
   - Device services now returns a http code 408 when anavailable [#370](https://github.com/cartoway/planner-web/pull/370)
+  - Route color reset correctly update stop popup "send_to_route" color [#381](https://github.com/cartoway/planner-web/pull/381)
 
 ## V105.0.0
   ### Added